+++ .github/workflows/ci.yaml
@@ -39,7 +39,6 @@ jobs:
     strategy:
       matrix:
         python-version: ["3.9", "3.10", "3.11", "3.12"]
-        pytest-markers: ["", "postgres"]
 
     steps:
       - name: Checkout code
@@ -63,7 +62,7 @@ jobs:
           pytest \
             --cov=autosubmit --cov-config=.coveragerc \
             --cov-report=xml:test/coverage.xml --cov-append \
-            test/unit -m "${{ matrix.pytest-markers }}"
+            test/unit -m ""
 
       - name: Coverage report
         run: |
@@ -73,7 +72,7 @@ jobs:
       - name: Upload coverage artifact
         uses: actions/upload-artifact@v4
         with:
-          name: coverage_unit_py-${{ matrix.python-version }}-${{ matrix.pytest-markers }}
+          name: coverage_unit_py-${{ matrix.python-version }}
           path: coverage.xml
           retention-days: 7
 
@@ -84,6 +83,7 @@ jobs:
     strategy:
       matrix:
         python-version: ["3.9", "3.10", "3.11", "3.12"]
+        pytest-markers: ["", "postgres"]
 
     steps:
       - name: Checkout code
@@ -111,7 +111,7 @@ jobs:
             --cov=autosubmit --cov-config=.coveragerc \
             --cov-report=xml:test/coverage.xml --cov-append \
             test/integration \
-            -m ""
+            -m "${{ matrix.pytest-markers }}"
 
       - name: Coverage report
         run: |
@@ -121,7 +121,7 @@ jobs:
       - name: Upload coverage artifact
         uses: actions/upload-artifact@v4
         with:
-          name: coverage_integration_py-${{ matrix.python-version }}
+          name: coverage_integration_py-${{ matrix.python-version }}-${{ matrix.pytest-markers }}
           path: coverage.xml
           retention-days: 7
