+++ pytest.ini
@@ -4,7 +4,7 @@ addopts =
     --strict-markers
     --doctest-modules
     -n auto --max-worker-restart 0 --dist loadgroup
-    -m "not docker"
+    -m "not docker and not postgres"
 testpaths =
     test/unit
     autosubmit
