+++ test/integration/database/test_db_manager.py
@@ -15,49 +15,36 @@
 # You should have received a copy of the GNU General Public License
 # along with Autosubmit.  If not, see <http://www.gnu.org/licenses/>.
 
+import pytest
+
 from autosubmit.database.db_manager import DbManager
 
 
-def test_create_table_command_returns_a_valid_command():
-    # arrange
-    table_name = 'tests'
-    table_fields = ['dummy1', 'dummy2', 'dummy3']
-    expected_command = 'CREATE TABLE IF NOT EXISTS tests (dummy1, dummy2, dummy3)'
-    # act
-    command = DbManager.generate_create_table_command(table_name, table_fields)
-    # assert
-    assert expected_command == command
+@pytest.mark.skip('It uses real data.')
+class TestDbManager:
+    def setup_method(self):
+        self.db_manager = DbManager('', 'test-db', 1)
 
+    def teardown_method(self):
+        self.db_manager.drop()
 
-def test_insert_command_returns_a_valid_command():
-    # arrange
-    table_name = 'tests'
-    columns = ['col1, col2, col3']
-    values = ['dummy1', 'dummy2', 'dummy3']
-    expected_command = 'INSERT INTO tests(col1, col2, col3) VALUES ("dummy1", "dummy2", "dummy3")'
-    # act
-    command = DbManager.generate_insert_command(table_name, columns, values)
-    # assert
-    assert expected_command == command
+    def test_db_manager_has_made_correct_initialization(self):
+        name = self.db_manager.select_first_where('db_options', ['option_name="name"'])[1]
+        version = self.db_manager.select_first_where('db_options', ['option_name="version"'])[1]
+        assert self.db_manager.db_name == name
+        assert self.db_manager.db_version == int(version)
 
+    def test_after_create_table_command_then_it_returns_0_rows(self):
+        table_name = 'test'
+        self.db_manager.create_table(table_name, ['field1', 'field2'])
+        count = self.db_manager.count(table_name)
+        assert 0 == count
 
-def test_insert_many_command_returns_a_valid_command():
-    # arrange
-    table_name = 'tests'
-    num_of_values = 3
-    expected_command = 'INSERT INTO tests VALUES (?,?,?)'
-    # act
-    command = DbManager.generate_insert_many_command(table_name, num_of_values)
-    # assert
-    assert expected_command == command
-
-
-def test_select_command_returns_a_valid_command():
-    # arrange
-    table_name = 'tests'
-    where = ['test=True', 'debug=True']
-    expected_command = 'SELECT * FROM tests WHERE test=True AND debug=True'
-    # act
-    command = DbManager.generate_select_command(table_name, where)
-    # assert
-    assert expected_command == command
+    def test_after_3_inserts_into_a_table_then_it_has_3_rows(self):
+        table_name = 'test'
+        columns = ['field1', 'field2']
+        self.db_manager.create_table(table_name, columns)
+        for i in range(3):
+            self.db_manager.insert(table_name, columns, ['dummy', 'dummy'])
+        count = self.db_manager.count(table_name)
+        assert 3 == count
